Added user authentication with passport so users can log in from the database securely.
