+++ cracking_the_resume/routes/index.js
@@ -16,102 +16,14 @@ router.get('/', function(req, res, next) {
     res.render('homePage', { title: 'home page',UsernameError:req.flash('UsernameError'), PasswordError:req.flash('PasswordError')});
 });
 
-// passport.use('local.signup', new local_strategy({
-//     usernameField: 'username',
-//     passwordField: 'password',
-//     passReqToCallback: true
-// }, function(req, username, password, done) {
-//     process.nextTick(function() {
-//     console.log("I am here");
-//
-//     User.findOne({'UserName':username}, function(err,user) {
-//         console.log("I am inside here");
-//
-//         if(err) {
-//             return done(err);
-//         }
-//
-//         if(user) {
-//             console.log("Username not found");
-//             return done(null, false);
-//         }
-//         else {
-//
-//             var password1 = req.body.password;
-//             var ePassword = User.hashPassword(password1);
-//
-//             var newUser = new User();
-//             newUser.firstName = req.body.firstName;
-//             console.log(req.body.firstName);
-//             newUser.lastName = req.body.lastName;
-//             newUser.Email = req.body.email;
-//             newUser.UserName = req.body.userName;
-//             newUser.Password = ePassword;
-//             newUser.Major = req.body.major;
-//             newUser.Seeking = req.body.seeking;
-//
-//             newUser.save(function (err) {
-//                 if (err) {
-//                     return done(err);
-//                 }
-//                 console.log("created a user");
-//                 return done(null, newUser);
-//             })
-//         }
-//
-//     })
-//     })
-//
-// }));
-//
-//
-// router.get('/signUp', function(req,res) {
-//     res.render('signUp');
-// });
-//
-
-//
-// router.post('/signUp',
-//     passport.authenticate('local.signup',
-//         {successRedirect: 'split',
-//             failureRedirect: '/',
-//             failureFlash: true}
-//     ));
-
-
-
-// router.post('/split', function(req, res, next) {
-//     var password1 = req.body.password;
-//     var ePassword = User.hashPassword(password1);
-//
-//     var newUser = new User();
-//     newUser.firstName = req.body.firstName;
-//     newUser.lastName = req.body.lastName;
-//     newUser.Email = req.body.email;
-//     newUser.UserName = req.body.userName;
-//     newUser.Password = ePassword;
-//     newUser.Major = req.body.major;
-//     newUser.Seeking = req.body.seeking;
-//
-//     newUser.save(function(err) {
-//         if(err) {
-//             return err;
-//         }
-//         console.log("created a user");
-//         return newUser;
-//     })
-//
-// });
-
-
-
-
+//Create a local strategy using passport every time a user needs to log in to page
 passport.use('local.login',new local_strategy({
     usernameField: 'username',
     passwordField: 'password',
     passReqToCallback: true
 }, function(req, username, password,done) {
 
+    //Find the user in the database
     User.findOne({'UserName': username}, function(err, user) {
 
         if(err) {
