+++ cracking_the_resume/routes/signup.js
@@ -1,8 +1,12 @@
 var express = require('express');
 var router = express.Router();
 
+var User = require('../Models/user');
+
+
 router.get('/', function(req, res, next) {
     res.render('signUp', { title: 'signup page' });
 });
+
 
 module.exports = router;
