+++ cracking_the_resume/views/account.ejs
@@ -27,6 +27,7 @@ fclose ($handle);}
   </head>
 
   <body>
+   <h2> <%=user.UserName%></h2>
     <nav class="navbar navbar-inverse navbar-static-top">
       <div class="container">
         <div class="navbar-header">
