+++ cracking_the_resume/views/homePage.ejs
@@ -9,6 +9,10 @@
   <script src="../routes/index.js"></script>
 </head>
 <body>
+
+     <h2> <%=UsernameError%></h2>
+     <h2> <%=PasswordError%></h2>
+
   <div style="" class="row theWhole">
     <div class="col-md-6 leftSide">
       <!-- <img class="img-responsive leftPic" src="https://img.etsystatic.com/il/48a2c6/914224541/il_fullxfull.914224541_ci1e.jpg" alt=""> -->
@@ -19,15 +23,15 @@
         <div class="col-md-3">
         </div>
         <div style="text-align:center;" class="col-md-6">
-          <form class="formRight">
+          <form class="formRight" method = "post" action="/">
             <div class="form-group">
              <label for="username"></label>
-             <input type="username" class="form-control" id="username1" placeholder="Enter username">
+             <input type="username" class="form-control" id="username1" placeholder="Enter username" name = "username">
             </div>
 
             <div class="form-group">
              <label for="password"></label>
-             <input type="password" class="form-control" id="password1" placeholder="Enter password">
+             <input type="password" class="form-control" id="password1" placeholder="Enter password" name = "password">
             </div>
 
             <button class="btn btn-login">
