+++ cracking_the_resume/views/redirect.ejs
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+
+<div style="text-align:center;" class="col-md-6">
+    <h1>Congratulations you have signed up for Cracking The Resume</h1>
+    <h1>Now please Log In</h1>
+    <a href = "/" class = "btn btn-signup btn-lg" role = "button"> Log In </a>
+</div>
+
+</html>
